Integrate search_state class into the search program
Before this patch, the search program was passed
a fitness function (called eval_sys_config) to score each system config.
Because we need to perserve the stats that are generated by the fitness
function, I created the SearchState class.  Calls to eval_sys_config()
are replaced by calls to search_state.eval_fitness().  At the
end of the search, the final system configs found in search.sys_configs
are used to access their associated stats in the search_state object
via the following:

for config in search.sys_configs:
    search_state.generate_job_output(config)