First prototype for the search monitor.
This branch is for development of an interactive search monitor. The user should
be able to interactively monitor the status of each search party in the search program.
This monitor will allow the user to:

1. View the sys_config and stats of the current best node for each search party

2. Terminate (or restart) a search party when it is searching an undesirable region

This commit adds a proof of concept design which will be tested and refined
before the final implementation of the search monitor