Added comments for clarity and fixed some typos
The stats are stored in a 2 layer dict as follows:

    mock_sim_stats = stats[sys_config][MockSim]

This entry, mock_sim_stats, contains a dictionary of statistics
from the run of the MockSim using sys_config as the system configuration

When checking constraints, if the current search_state contains
more than one simulation wrapper, the various stats must all be
collected and checked. For example:

    mock_sim_stats = stats[sys_config][MockSim]
    gem_5_stats = stats[sys_config][Gem5Sim]

Both mock_sim_stats and gem_5_stats must be checked for constraint
violations.