Implemented McPAT search state
This commit implements a McPAT search state and includes a test that
runs a search using only the McPAT simulator.

Additionally, I changed how McPAT validates parameters, so the whole
system config can be passed to the simulator.  Before all the given
sys config parameters needed to be valid for McPAT, but this method
requires the search state to specific pick out the parameters needed
for McPAT.  Now, McPAT takes all the sys_config parameters and only
uses the ones that it finds valid.  An error is only thrown if none of
the given parameters are valid for McPAT.