Cleaned up McPatSim init function.
We don't need to pass a system_config on initialization and we're
currently just passing a empty sys config which won't work anyway.

This commit removes the sys_config parameter from the McPatSim
constructor.  It also modifies the McPatSim search state and tests to
reflect this change to the McPatSim.

We should do a similar change for the mock sim and the gem5 sim.