Added plumbing for synchrotrace wrapper.
Instead of a benchmark, the synchrotrace wrapper takes a event trace
that is generated by the Sigil2 tool.  Integration of the Sigil2 tool
into the synchrotrace wrapper will take place in a later commit.

Currently, the syntrace_wrapper wrapper can run a basic simulation
with a single threaded trace and all parameters are hard coded.
Further commits will allow modification of the syncrhotrace parameters
by the user.

Note:  A trace contains information about the number of cores, so a
new trace will be needed every time we adjust the number of CPU cores.